Cloud_funcations/Fetch_Data_aqicn/requirements.txt
flask==2.0.1
gunicorn==20.1.0
google-cloud-storage==2.13.0
python-dotenv==1.0.0
requests==2.31.0
